proyectoCompilador/src/requirements.txt
antlr4-python3-runtime==4.13.1
flet